Move IssueUser code to separate file
Also add unit tests
